+++ ext/Shared/Config.lua
@@ -100,7 +100,6 @@ Config = {
 	MaxShootDistanceNoAntiAir = 175,	-- Meters bots in vehicle (no Anti-Air) starts shooting at players 
 	VehicleWaitForPassengersTime = 7.0,	-- Seconds to wait for other passengers 
 	ChopperDriversAttack = false,		-- If false, choppers only attack without gunner on board 
-	VehicleMoveWhileShooting = true,	-- Vehicles like tanks do not stop for shooting 
 	AABots = false,						-- Enable Auto-AA by NyScorpy 
 	MaxDistanceAABots = 375,			-- Max Range of Stationary AA 
 	EnableParadrop = false,				-- Bots can spawn on vehicles inside C-130 gunship