+++ .github/CHANGELOG.md
@@ -22,6 +22,7 @@ This is the changelog for the version V3.0 Don't forget to [join us on Discord](
 * Bots can use every jet in conquest now (no more paths needed)
 * No more air-paths needed (only enter-paths for choppers)
 * Kind of Support for AirSuperiority (jets still too passive)
+* Bots randomly move or stop while shooting in vehicles (by MatiasPastori)
 
 ### some open TODOs:
 * rework of raycasts for better performance